Changing app, testing deploy

+++ deployment/scripts/test.sh
@@ -10,6 +10,9 @@ docker compose -f ./deployment/docker-compose.yaml up  -d
 
 docker exec app pytest
 
+failure=`docker exec app pytest | grep 'FAILED' | awk '{print $1}'`
+if [ -z "$failure" ]; then echo "passed"; else exit 1; fi
+
 docker rm -f $(docker ps -aq)
 
 rm ./deployment/.env
